requirements.txt
tqdm==4.65.0
requests==2.31.0
Pillow==9.3.0